requirements.txt
requests
cryptography
wheel
tlslite-ng==0.8.0a43
srp
pbkdf2
unicorn
rich
prompt_toolkit